to-jsonschema: Fix pattern keyword for strings
JSON Schema supports the `pattern` keyword to express constraints for
strings as regular expressions [1].

Update the to-jsonschema module to return `pattern` as a string instead
of as a RegExp object. This ensures a valid `pattern` in the generated
JSON Schema.

Note that JSON Schema doesn't support regular expression flags in its
patterns.

[1] https://json-schema.org/understanding-json-schema/reference/regular_expressions.html